Removed development config, default and testing values are in default
