+++ config/index.js
@@ -1,5 +1,4 @@
 var _ = require("lodash");
 var defaults = require("./default.js");
-var config = require("./" + (process.env.NODE_ENV || "development") + ".js");
+var config = require("./" + (process.env.NODE_ENV || "default") + ".js");
 module.exports = _.merge({}, defaults, config);
-
